# server/openapi.yaml
openapi: 3.0.3
info:
  title: OpenBin API
  version: 1.0.0
  description: |
    REST API for OpenBin — a multi-user web app for organizing physical storage bins with QR codes.

    **Authentication**: Most endpoints require authentication via one of:
    - **httpOnly cookies** (primary) — `openbin-access` (15 min) + `openbin-refresh` (7 days), set automatically on login/register. Browsers send these with every request.
    - **API key** created via `POST /api-keys` (long-lived, per-user, format: `sk_openbin_...`) — `Authorization: Bearer <key>`. A key grants access to all locations the user belongs to.

    **Response envelopes**: All list endpoints return `{ results: T[], count: number }`.
    Error responses use `{ error: "CODE", message: "Human readable" }`.

    **Rate limiting**: All API endpoints share a global limit of 200 requests/minute.
servers:
  - url: /api
    description: API base path (proxied via nginx)
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: API key (sk_openbin_...) via Authorization header. Cookies are the primary auth method for browsers.
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          enum:
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - NOT_FOUND
            - FORBIDDEN
            - CONFLICT
            - RATE_LIMITED
            - INTERNAL_ERROR
            - GONE
            - MALWARE_DETECTED
            - SCAN_UNAVAILABLE
            - SERVICE_NOT_CONFIGURED
        message:
          type: string
      example:
        error: NOT_FOUND
        message: Bin not found

    # ── Auth ───────────────────────────────────────────────
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
        email:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        plan:
          type: string
          enum: [lite, pro]
          description: Current plan tier
        subscriptionStatus:
          type: string
          enum: [inactive, active, trial]
          description: Current subscription status
        activeUntil:
          type: string
          format: date-time
          nullable: true
          description: Date until which the current plan is active (null for self-hosted)
        language:
          type: string
          nullable: true
          description: >
            User's explicit language preference (e.g. "en", "nb"), set only via
            Settings → Preferences. Null means no explicit preference — the
            client falls back to browser-detected/localStorage language.
...
    UpdateProfileRequest:
      type: object
      properties:
        displayName:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
        language:
          type: string
          nullable: true
          description: Must be one of the server's supported language codes (currently "en", "nb").
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
          description: Must contain uppercase, lowercase, and digit
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: UNAUTHORIZED
            message: Not authenticated
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: NOT_FOUND
            message: Resource not found
    ValidationError:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: VALIDATION_ERROR
            message: Name is required
...
paths:
...
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user profile
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/User'
                  - type: object
                    properties:
                      activeLocationId:
                        type: string
                        format: uuid
                        nullable: true
                        description: Persisted active location selection (null if not set or membership lost)
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/active-location:
    put:
      tags: [Auth]
      summary: Persist active location selection
      description: Sets the user's active location. The location is validated for membership.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [locationId]
              properties:
                locationId:
                  type: string
                  format: uuid
                  nullable: true
                  description: Location ID to set as active, or null to clear
      responses:
        '200':
          description: Active location updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  activeLocationId:
                    type: string
                    format: uuid
                    nullable: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Not a member of the specified location
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/profile:
    put:
      tags: [Auth]
      summary: Update display name, email, and/or language
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/password:
    put:
      tags: [Auth]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password updated
        '401':
          description: Current password incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /auth/avatar/{userId}:
    get:
      tags: [Auth]
      summary: Serve avatar file
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Avatar image file
          content:
            image/*:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...